+++ ChangeLog.md
@@ -10,6 +10,7 @@
     - Define the set of CRUD actions, alogn with a allowFn() configuration parameter, and a Roles.isAllowed() new function, bumping minor candidate version number
     - Define Roles.suggestedPermissions() new function
     - Define Roles.compareLevels() new function
+    - Reinit scopes before reloading them
 
 ### 1.5.0